noitification details and capital first word and bug resolved of rating and doc is added in the model

+++ app/src/main/java/com/carrus/carrusshipper/model/Rating.java
@@ -10,10 +10,10 @@ import java.io.Serializable;
 public class Rating  implements Serializable {
 
     @SerializedName("shipperRate")
-    public Long shipperRate;
+    public Double shipperRate;
 
     @SerializedName("truckerRate")
-    public Long truckerRate;
+    public Double truckerRate;
 
     @SerializedName("shipperComment")
     public String shipperComment;
